1. 代码生成器，配置路径使用类资源加载类 2. 修改字段长度size的类型,int=>long

+++ mldong-generator/src/main/java/com/mldong/generator/core/impl/MysqlDataBase.java
@@ -173,8 +173,8 @@ public class MysqlDataBase implements DataBase{
 			// 表名
 			// String tableName = rs.getString("TABLE_NAME");
 			column.setColumnName(columnName);
-			int size = rs.getInt("CHARACTER_MAXIMUM_LENGTH");
-			column.setSize(size==0?rs.getInt("NUMERIC_PRECISION"):size);
+			long size = rs.getLong("CHARACTER_MAXIMUM_LENGTH");
+			column.setSize(size==0L?rs.getLong("NUMERIC_PRECISION"):size);
 			column.setNullable("YES".equals(rs.getString("IS_NULLABLE")));
 			column.setDefaultValue(rs.getString("COLUMN_DEFAULT"));
 			column.setDataType(rs.getString("DATA_TYPE"));
